Chat/requirements.txt
torch
transformers
fastapi
uvicorn
google-cloud-aiplatform
vertexai.preview.language_models
